+++ android/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="mmorihiro.larger_circle" >
+          package="mmorihiro.matchland">
 
     <uses-sdk android:minSdkVersion="19" android:targetSdkVersion="26" />
 
@@ -11,9 +11,9 @@
         android:label="@string/app_name"
         android:theme="@style/GdxTheme" >
         <activity
-            android:name="mmorihiro.larger_circle.android.AndroidLauncher"
-            android:label="@string/app_name"
-            android:configChanges="keyboard|keyboardHidden|orientation|screenSize">
+                android:name="mmorihiro.matchland.android.AndroidLauncher"
+                android:label="@string/app_name"
+                android:configChanges="keyboard|keyboardHidden|orientation|screenSize">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
