--- core/src/mmorihiro/larger_circle/controller/Controller.kt
@@ -1,8 +1,0 @@
-package mmorihiro.larger_circle.controller
-
-import mmorihiro.larger_circle.view.View
-
-
-interface Controller {
-    val view: View
-}
