--- core/src/mmorihiro/larger_circle/view/ConnectEvent.kt
@@ -1,14 +1,0 @@
-package mmorihiro.larger_circle.view
-
-import mmorihiro.larger_circle.model.Point
-
-data class ConnectEvent(val connectedItems: List<Point>,
-                        val sameTypeGroup: Set<Point>,
-                        val enemy: List<Point>) {
-    init {
-        require(connectedItems.isNotEmpty())
-        require(sameTypeGroup.isNotEmpty())
-    }
-
-    val enemyPoint = enemy.last()
-}
