--- core/src/mmorihiro/larger_circle/view/MyImage.kt
@@ -1,6 +1,0 @@
-package mmorihiro.larger_circle.view
-
-import com.badlogic.gdx.scenes.scene2d.ui.Image
-
-
-class MyImage(image: Image, var type: Pair<Int, Int>) : Image(image.drawable)
