+++ core/src/mmorihiro/matchland/controller/Controller.kt
@@ -1,0 +1,8 @@
+package mmorihiro.matchland.controller
+
+import mmorihiro.matchland.view.View
+
+
+interface Controller {
+    val view: View
+}
