+++ core/src/mmorihiro/matchland/controller/Extensions.kt
@@ -1,0 +1,7 @@
+package mmorihiro.matchland.controller
+
+import com.badlogic.gdx.math.Circle
+import com.badlogic.gdx.scenes.scene2d.Actor
+
+
+fun Actor.circle(): Circle = Circle(x + width / 2, y + height / 2, width / 2)
