+++ core/src/mmorihiro/matchland/view/MyImage.kt
@@ -1,0 +1,6 @@
+package mmorihiro.matchland.view
+
+import com.badlogic.gdx.scenes.scene2d.ui.Image
+
+
+class MyImage(image: Image, var type: Pair<Int, Int>) : Image(image.drawable)
