+++ core/src/mmorihiro/matchland/view/StageChangeView.kt
@@ -1,0 +1,23 @@
+package mmorihiro.matchland.view
+
+import com.badlogic.gdx.graphics.Color
+import ktx.actors.centerPosition
+import ktx.scene2d.KTextButton
+import ktx.scene2d.textButton
+import ktx.scene2d.window
+import mmorihiro.matchland.model.Values
+import kotlin.properties.Delegates
+
+
+class StageChangeView : View() {
+    var button by Delegates.notNull<KTextButton>()
+        private set
+    val window = window(title = "Stage Completed!") {
+        defaults().pad(20f)
+        button = textButton("continue") {
+            color = Color.GREEN
+        }
+        pack()
+        centerPosition(Values.width, Values.height)
+    }
+}
