--- model/src/main/mmorihiro/larger_circle/model/Values.kt
@@ -1,8 +1,0 @@
-package mmorihiro.larger_circle.model
-
-object Values {
-    const val width = 288f
-    const val height = 480f
-    const val centerX = width / 2
-    const val centerY = height / 2
-}
